updates to int64 for stats

+++ pkg/storage/unified/search/dashboard_stats.go
@@ -10,6 +10,6 @@ func ProvideDashboardStats() *dashboardStats {
 	return &dashboardStats{}
 }
 
-func (s *dashboardStats) GetStats(ctx context.Context, namespace string) (map[string]map[string]int, error) {
+func (s *dashboardStats) GetStats(ctx context.Context, namespace string) (map[string]map[string]int64, error) {
 	return nil, nil
 }
